pyproject.toml
[project]
name = "external-models-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.80.11",
    "mcp>=1.25.0",
    "python-dotenv>=1.2.1",
]
